==> notebooks/4.0-semEval22-seql-dev-eda.ipynb <==
## SemEval 2022 Dev set EDA: Seq-labeling Baseline
- Author: Nana
- Created: 2022.7.19

- Dataset statistics 
    src: [1.0-semEval22-eda.ipynb](https://github.com/Nana2929/ckip_absa_en/blob/main/notebooks/1.0-semEval22-eda.ipynb)
    - 資料數量
    
    |-|opener_en|darmstadt_unis|mpqa|
    |-------|------|------|-------|
    |train  |1744  |2253  |5643   |
    |dev    |249   |232   |1997   |
    |test   |499   |318   |2055   |

    - 無情緒標記資料數量
    
    |-|opener_en|darmstadt_unis|mpqa|
    |-------|------|------|-------|
    |train  |344   |1572  |4389   |
    |dev    |51    |150   |1581   |
    |test   |92    |214   |1665   |

The below description is taken from [seq-labeling Github](https://github.com/jerbarnes/semeval22_structured_sentiment/tree/master/baselines/sequence_labeling).
- Pipelined; first train **3 separate models that extract the 3 sub-elements (holders, targets, expressions)** using sequence labeler (BiLSTM, 1-layer, 100 hidden-dim) for 10 epochs
- Finally, train a **relation prediction model (BiLSTM + max-pooling)** to create contextualized representations of 1) the full text, 2) the first element (either a holder or target) and 3) the sentiment expression.
- These three representations are then concatenated and passed to a linear layer followed by a sigmoid function. The training consists of predicting whether two elements have a relationship or not, converting the problem in binary classification.

- Each dataset is trained alone (**7 datasets * (3+1) models = 28 models**).
- The best development scores (metrics: **Macro F1**):
1. extractions 直接看bios tagging每個token label正確與否
2. relations (target, expression)兩兩交叉配對，有關係就T無關係則F，binary classifcation 

|annotation_type|  opener_en |darmstadt_unis|  mpqa  |
|-----------|-----------|-------------|--------|
|sources    |     0.779   |      0.547     |      0.432       |
|targets    |     0.726    |     0.459    |         0.511    |
|expressions|   0.577      |   0.243      |       0.320     | 
|relations|     0.701   | 0.435        |      0.709       | 


import os
os.chdir('/share/home/nana2929/')

import json
root_dir = './semEval22/Structured_Sentiment/'
prediction_dir = f'{root_dir}/baselines/sequence_labeling/saved_models/relation_prediction'
golden_dir = f'{root_dir}/data'
PREDICTIONS = {'darmstadt_unis':None, 'mpqa':None, 'opener_en':None}
GOLDS = {'darmstadt_unis':None, 'mpqa':None, 'opener_en':None}

for DATASET in PREDICTIONS.keys():
    predpath = f'{prediction_dir}/{DATASET}/prediction.json'
    goldpath = f'{golden_dir}/{DATASET}/dev.json'
    with open (predpath, 'r') as f:
        PREDICTIONS[DATASET] = json.load(f)  
    with open (goldpath, 'r') as f2:
        GOLDS[DATASET] = json.load(f2) 

ALLDEVS = {}

# match by sent-id
from collections import defaultdict
def pred_align(preds, golds):
    '''
    preds: list of dicts
    golds: list of dicts
    dict內需包含 'sent_id' key 
    '''
    SENTID = defaultdict(lambda: {'PRED':None, 'GOLD':None})
    for p, g in zip(preds, golds):
        p_id, g_id = p['sent_id'], g['sent_id']
        SENTID[p_id]['PRED'] = p
        SENTID[g_id]['GOLD'] = g
    return SENTID, list(SENTID.keys())

def show_random(dataset_name, sent_id = None):
    '''
    dataset_name: string
    '''
    from random import randint
    try:
        PGPAIRS, ids = ALLDEVS[dataset_name]
    except KeyError:
        print(f'No {dataset_name} in ALLDEVS.')
    dev_size=len(ids)
    rsid = ids[randint(0, dev_size-1)]
    if sent_id:
        rsid = sent_id
    
    print(f'[INFO] dev set size: {dev_size}')
    print(f'[INFO] sent id: {rsid}')
    pred, gold = PGPAIRS[rsid]['PRED'], PGPAIRS[rsid]['GOLD']
    return pred, gold, rsid

keys = ('opener_en', 'darmstadt_unis', 'mpqa')

for key in keys:
    PGPAIRS, ids = pred_align(PREDICTIONS[key], GOLDS[key])
    ALLDEVS[key] = (PGPAIRS, ids)

### A. opener_en

#### example 1: short

key = 'opener_en'
pred, gold, sent_id = show_random(key, sent_id='opener_en/kaf/hotel/english00122_b509ee1b998f3418b1fd4a8096880441-1')

gold

pred

#### example 2: multiple, simple 
determine relation的部分似乎配對得不太好，把所有extractions都交叉配對了，
但這筆資料extractions近乎全對

- text = 'the hotel is very clean and all the people who are working their are very friendly and willing to help .'

整體prediction
- the hotel: clean, friendly, willing to help
- the people: clean, friendly, willing to help

pred, gold, sent_id = show_random(key)

gold

pred

#### example 3: multiple, complicated expressions
expressions 只能抓到 'a little rough from all of the Hen'，忽視了cleanest
targets 的部分 reception 沒有抓出來，只有抓出定冠詞。

pred, gold, sent_id = show_random(key)

gold

pred

#### example 4: implicit sentiment 

40歐元顯然是說太貴的意思，
但這個如果沒有world knowledge是不能判斷的（模型預測無標記）

pred, gold, sent_id = show_random(key)
gold

pred

### B. darmstadt_unis

key = 'darmstadt_unis'

#### example 1: short

- gold: quality: poor
- pred: they: poor, isbn: poor

extractions沒有全對，沒有抓出quality這個字（抓出了they, isbn），因為pipeline的error propagation，
relation pairing也是錯的。expression和情緒極度是對的

pred, gold, sent_id = show_random(key)

gold

pred

#### example 2: short, 'joke'


Capella_University_80_09-21-2004-1 

unregular expression? 抓得還算可以，但target標得不完善

是說這個資料集，很多都是空opinions（pred和gold都是）

pred, gold, sent_id = show_random(key) 

gold

pred

#### example 3: Neutral和不標的差異到底在...？

pred, gold, sent_id = show_random(key)

pred

gold

### C. mpqa

key = 'mpqa'

#### example 1: weird prediction
determination被標成negative


pred, gold, sent_id = show_random(key)

gold

pred

#### example 2: non-empty, all-correct
看起來完全正確的標記，’The SADC on Saturday endorsed the vote' 
extractions全對，因為只有一組，relation pairing is correct (trivially)
唯有intensity略有出入但影響不大

pred, gold, sent_id = show_random(key, sent_id='20020316/20.37.48-18053-9')

gold

pred

#### example 3: 6 opinions in one data, complicated relations and target spans
這個資料集的targets的定義真的很弔詭，常常是phrases（boundary會很難抓），而且polar expressions有動詞
這個資料的relation pairing prediction很慘，不過他似乎有學到 n't 是否定的意思。

pred, gold, sent_id = show_random(key, sent_id='ula/sw2015-ms98-a-trans-23')

gold

pred
