==> notebooks/32.4-yelp-inference-format.ipynb <==
import json 
import re
from copy import deepcopy

DATA = '../data/yelp'
f1p = f'{DATA}/RAW_yelp-food-test_2rest100.json'
f2p = f'{DATA}/RAW_yelp-food-test_rand100.json'
with open(f1p, 'r') as f:
    f1 = json.load(f)
with open(f2p, 'r') as f:
    f2 = json.load(f)

## Transformation Function

# https://stackoverflow.com/questions/44244583/splitting-on-regex-without-removing-delimiters
def transform(datas):
    '''
    input: datas: a list of json dictionary 
    return: 
    '''
    RESULT = []
    print(f'Input Data Size: {len(datas)}')
    for data in datas:
        review_id = data['review_id']
        text = data['text']
        # cleaning text
        # text = re.sub('/n', '', text) 
        # # separating
        # text = re.sub('[.|!|?]', '|', text)
        # sentences = text.split('|')
        sentences = re.split(r'(?<=[\.\!\?])\s*', text)
        idx = 0 
        for sentence in sentences:
            sentence = sentence.strip()
            # print(review_id, idx)
            if len(sentence) == 0:
                continue
            this_data = deepcopy(data)
            this_data['sent_id'] = f'{review_id}:{idx}'
            this_data['text'] = sentence
            this_data['opinions'] = []
            # print(this_data)
            
            RESULT.append(this_data)
            idx += 1 
            
    print(f'Total Size After Transform: {len(RESULT)}')
    return RESULT
            
f1_ = transform(f1)        

f2_ = transform(f2) 

import random 
x = f2_
i = random.randint(0, len(x))
x[i]

## Output

new_f1p = f1p.lstrip('RAW_')
new_f2p = f2p.lstrip('RAW_')


new_fp = [new_f1p, new_f2p]
new_files = [f1_, f2_]
for new_con, fp in zip(new_files, new_fp):
    with open(fp, 'w') as f:
        json.dump(f, indent = 4, ensure_ascii = False)
