==> notebooks/33.0-s16p-opener-merge.ipynb <==
import json

with open('../../../semeval22_structured_sentiment/data/opener_en/train.json') as f:
    oper = json.load(f)

len(oper)

# count number of opinions

count = 0
for i in oper:
    count += len(i['opinions'])

count

# add key of dataset

for i in range(len(oper)):
    oper[i]['dataset'] = 'opener_en'

oper[:10]

with open('../data/semEval16_plus/s16+_train.json') as f:
    s16p = json.load(f)

len(s16p)

# count number of opinions

count = 0
for i in s16p:
    count += len(i['opinions'])

count

merge_list = oper

for i in s16p:
    merge_list.append(i)

len(merge_list)

# output to json file

with open('../semEval16_plus/mixed_train.json', 'w') as outfile:
    json.dump(merge_list, outfile)
