==> notebooks/35.0-polar-stars-2rest100.ipynb <==
### Check the polarity of inference results based on stars of each review
##### data: 100 random-selected reviews from 2 restaurants

import pandas as pd
import json

# open the inference results

with open('../data/yelp/INFER_yelp-food-test_2rest100.conllu.json') as f:
    infer = json.load(f)

infer[:5]

# open the raw reviews

with open('../data/yelp/yelp-food-test_2rest100.json') as f:
    raw = json.load(f)

raw[:5]

# add some information (stars, business_id,...) to inferrence from raw reviews

for i in range(len(infer)):
    raw_one = [j for j in raw if j['sent_id'] == infer[i]['sent_id']][0]
    # print(infer[i]['sent_id'], raw_one['sent_id'])
    infer[i]['business_id'] = raw_one['business_id']
    infer[i]['review_id'] = raw_one['review_id']
    infer[i]['stars'] = raw_one['stars']
    infer[i]['polar_ans'] = False

infer[:5]

# output json 
with open('../data/yelp/ANS-polar-stars_2rest100.json', 'w') as outfile:
    json.dump(infer, outfile, indent = 4, ensure_ascii=False)

# count the number of non-empty opinions
c = 0
for i in infer:
    if len(i['opinions']) > 0:
        c += 1

print(f'total: {len(infer)}')
print(f'non-empty: {c}')

##### EDA of annotated 2rest100 (after annotation on the above output json)

with open('../data/yelp/ANS-polar-stars_2rest100.json') as f:
    infer_gold = json.load(f)

infer_gold[:5]

# restaurant1: '_ab50qdWOk0DdB6XOrBitw'
# restaurant2: 'ac1AeYqs8Z4_e2X5M3if2A' 

empty = 0
rest1 = []
rest2 = []

for i in infer_gold:
    if len(i['opinions']) == 0:
        empty += 1
    else:
        if i['business_id'] == '_ab50qdWOk0DdB6XOrBitw':
            rest1.append(i)
        else:
            rest2.append(i)

print(f'non-empty: {len(infer_gold)-empty}, empty: {empty}')
print(f'r1: {len(rest1)}')
print(f'r2: {len(rest2)}')

# stars -> polar_ans -> True or Negative

print('---------------rest1----------------')

for k in range(2):

    s1, s2, s3, s4, s5 = 0, 0, 0, 0, 0 # Stars
    t1, t2, t3, t4, t5 = 0, 0, 0, 0, 0 # True
    tp1, tp2, tp3, tp4, tp5 = 0, 0, 0, 0, 0 # True Positive
    tn1, tn2, tn3, tn4, tn5 = 0, 0, 0, 0, 0 # True Negative

    if k == 0:
        lst = rest1
    else:
        lst = rest2

    for i in lst:
        if i['stars'] == 1.0:
            s1 += 1
            if i['polar_ans']:
                t1 += 1
                if i['opinions'][0]['Polarity'] == 'Positive':
                    tp1 += 1
                else:
                    tn1 += 1
        elif i['stars'] == 2.0:
            s2 += 1
            if i['polar_ans']:
                t2 += 1
                if i['opinions'][0]['Polarity'] == 'Positive':
                    tp2 += 1
                else:
                    tn2 += 1
        elif i['stars'] == 3.0:
            s3 += 1
            if i['polar_ans']:
                t3 += 1
                if i['opinions'][0]['Polarity'] == 'Positive':
                    tp3 += 1
                else:
                    tn3 += 1
        elif i['stars'] == 4.0:
            s4 += 1
            if i['polar_ans']:
                t4 += 1
                if i['opinions'][0]['Polarity'] == 'Positive':
                    tp4 += 1
                else:
                    tn4 += 1
        else:
            s5 += 1
            if i['polar_ans']:
                t5 += 1
                if i['opinions'][0]['Polarity'] == 'Positive':
                    tp5 += 1
                else:
                    tn5 += 1

    print(f'1s: {s1:3d}(T:{t1:2d}, F:{s1-t1:2d}) | (TP:{tp1:2d}, TN:{tn1:2d})')
    print(f'2s: {s2:3d}(T:{t2:2d}, F:{s2-t2:2d}) | (TP:{tp2:2d}, TN:{tn2:2d})')
    print(f'3s: {s3:3d}(T:{t3:2d}, F:{s3-t3:2d}) | (TP:{tp3:2d}, TN:{tn3:2d})')
    print(f'4s: {s4:3d}(T:{t4:2d}, F:{s4-t4:2d}) | (TP:{tp4:2d}, TN:{tn4:2d})')
    print(f'5s: {s5:3d}(T:{t5:2d}, F:{s5-t5:2d}) | (TP:{tp5:2d}, TN:{tn5:2d})')
    if k == 1:
        break
    print('---------------rest2----------------')
