==> notebooks/36.0-polar-major-2rest100.ipynb <==
import pandas as pd
import json

with open('../data/yelp/ANS_polar-stars_2rest100.json') as f:
    gold = json.load(f)

gold[:2]

# rest1: '_ab50qdWOk0DdB6XOrBitw'
# rest2: 'ac1AeYqs8Z4_e2X5M3if2A' 

# split into 2 rests
gold1, gold2 = [], []

for i in gold:
    if i['business_id'] == '_ab50qdWOk0DdB6XOrBitw':
        gold1.append(i)
    else:
        gold2.append(i)

print(len(gold), len(gold1), len(gold2))

import nltk
nltk.download('stopwords')
from nltk.corpus import stopwords

# define stopwords
stopword = stopwords.words('english') 
stopwords

import string
string.punctuation

# build a list of all target
target_list1, target_list2 = [], []

for goldlist in [gold1, gold2]:

    L = []

    for sent in range(len(goldlist)):
        opi = goldlist[sent]['opinions']

        if len(opi) > 0: # non empty opinion
            for i in range(len(opi)):
                goldlist[sent]['opinions'][i]['Target_key'] = [] # add new key 'Target_key' containing tokens of 'Target'
                # print(opi[i])

                if len(opi[i]['Target'][0]) > 0: # pred_target exists
                    for tgt in opi[i]['Target'][0]:
                        # remove all punctuation
                        tgt = tgt.translate(str.maketrans('', '', string.punctuation))

                        # split the target text
                        tgt_tokens = nltk.word_tokenize(tgt.lower())

                        # remove stopwords
                        tgt_tokens_nostop = [k for k in tgt_tokens if k not in stopword]
                        # print(tgt_tokens_nostop)

                        # add all words to the list
                        if len(tgt_tokens_nostop) > 0:
                            for t in tgt_tokens_nostop:
                                L.append(t)
                                goldlist[sent]['opinions'][i]['Target_key'].append(t)

    if goldlist == gold1:
        target_list1 = L
    else:
        target_list2 = L

gold1[:5]

# check different opinions' polarity of one sent when length of opinions > 1

for GOLD in [gold1, gold2]:
    r = 1
    print(f'------rest{r}------')
    r += 1
    
    diff = 0
    same = 0
    total = 0
    for i in GOLD:
        if len(i['opinions']) > 0:
            total += 1
        if len(i['opinions']) > 1:
            flag = 1

            for j in range(len(i['opinions'])):
                if (j != 0) and (i['opinions'][j]['Polarity'] != i['opinions'][j-1]['Polarity']):
                    flag = 0 # different polar
                    # print(j['Polarity'])
                    break
            
            if flag == 0:
                diff += 1
            else:
                same += 1

    print(f'total(opinion > 0): {total}\ndiff polar(opinion > 1): {diff}\nsame polar(opinion > 1): {same}')

# dict of word frequency
counts1, counts2 = {}, {}

for word in target_list1:
    counts1[word] = counts1.get(word,0) + 1

for word in target_list2:
    counts2[word] = counts2.get(word,0) + 1

print(counts1)
print('------------------')
print(counts2)

print(len(list(counts1.keys())), len(list(counts2.keys())))

# combine the word frquency of singular and plural noun
# leave singular noun only

for COUNTS in [counts1, counts2]:
   
    for k1 in list(COUNTS.keys()):
        for k2 in list(COUNTS.keys()):
            if (len(k1) < len(k2)) and (k2 == k1 + 's'):
                # print(k1, v1, k2, v2)
                COUNTS[k1] += COUNTS[k2]
                del COUNTS[k2]

print(counts1)
print('------------------')
print(counts2)    

print(len(list(counts1.keys())), len(list(counts2.keys())))

# Target Word Frequency Top 10

print('Target Word Frequency Top 10')

r = 1
for COUNTS in [counts1, counts2]:

    print(f'------rest{r}------')
    r += 1

    items = list(COUNTS.items())
    items.sort(key=lambda x:x[1], reverse=True)

    for j in range(10):
        word, count = items[j]
        print("{0:<10}{1:>5}".format(word, count))

# singular and plural form are consideres the same in 'Target_key'

for GLD in [gold1, gold2]:

    if GLD == gold1:
        KEY = list(counts1.keys())
    else:
        KEY = list(counts2.keys())
        
    for i in range(len(GLD)):
        if len(GLD[i]['opinions']) > 0: # non empty opinion
            
            for j in range(len(GLD[i]['opinions'])):
                if (len(GLD[i]['opinions'][j]['Target_key'])) > 0:

                    for t in range(len(GLD[i]['opinions'][j]['Target_key'])):
                        word = GLD[i]['opinions'][j]['Target_key'][t]
                        if (word.endswith('s')) and (word not in KEY) :
                            # print(word)
                            GLD[i]['opinions'][j]['Target_key'][t] = word[:-1] # remove 's' from plural words

gold1[:3]


import pandas as pd

# rest1's statistic results by target key
# df column: key | total_sent| TRUE_polar_ans | FALSE_polar_ans | accuracy | TP | TN | FP | FN | total_opinion | Positive | Negative
# important: TP, TN, FP and FN depend on first opinion's polarity only!!!

key1 = list(counts1.keys())
total_sent1, TRUE_polar_ans1, FALSE_polar_ans1, accur1 = [], [], [], []
total_opinion1, Pos1, Neg1, TP1, FP1 = [], [], [], [], []
TN1, FN1 = [], []

for k in key1:
    sent, t, f = 0, 0, 0
    opinion, pos, neg, tp, fp = 0, 0, 0, 0, 0

    for g in gold1:
        sign = False
        if (len(g['opinions']) > 0): # non empty 'opinions'
            for o in g['opinions']: # every opinion in 'opinions'
                if (len(o['Target_key']) > 0): # non empty 'Target_key'
                    for tgk in o['Target_key']: # every word in 'Target_key'

                        if tgk == k: 
                            sign = True
                            opinion += 1
                            if o['Polarity'] == 'Positive':
                                 pos += 1
                            else: 
                                neg += 1

            if sign == True:
                sent += 1
                if g['polar_ans']: # gold ans: True
                    t += 1
                    if g['opinions'][0]['Polarity'] == 'Positive': 
                        tp += 1
                else: # gold ans: False
                    f += 1
                    if g['opinions'][0]['Polarity'] == 'Positive': 
                        fp += 1
    
    total_sent1.append(sent)
    TRUE_polar_ans1.append(t)
    FALSE_polar_ans1.append(f)
    accur1.append("%.2f%%" % (100 * t/sent))
    total_opinion1.append(opinion)
    Pos1.append(pos)
    Neg1.append(neg)
    TP1.append(tp)
    FP1.append(fp)
    TN1.append(t-tp)
    FN1.append(f-fp)



# rest1's stat

df_stat1 = pd.DataFrame({'key': key1, 'total_sent': total_sent1, 'TRUE_polar_ans': TRUE_polar_ans1, 'FALSE_polar_ans': FALSE_polar_ans1,
                        'accuracy': accur1, 'TP': TP1, 'TN':  TN1, 'FP': FP1, 'FN': FN1,
                        'total_opinion': total_opinion1, 'Positive': Pos1, 'Negative': Neg1})

df_stat1 = df_stat1.sort_values(['total_sent', 'total_opinion'], ascending=False)
df_stat1[df_stat1.total_sent > 2]
# df_stat1

# rest2's statistic results by target key
# df column: key | total_sent| TRUE_polar_ans | FALSE_polar_ans | accuracy | TP | TN | FP | FN | total_opinion | Positive | Negative
# important: TP, TN, FP and FN depend on first opinion's polarity only!!!

key2 = list(counts2.keys())
total_sent2, TRUE_polar_ans2, FALSE_polar_ans2, accur2 = [], [], [], []
total_opinion2, Pos2, Neg2, TP2, FP2 = [], [], [], [], []
TN2, FN2 = [], []

for k in key2:
    sent, t, f = 0, 0, 0
    opinion, pos, neg, tp, fp = 0, 0, 0, 0, 0

    for g in gold2:
        sign = False
        if (len(g['opinions']) > 0): # non empty 'opinions'
            for o in g['opinions']: # every opinion in 'opinions'
                if (len(o['Target_key']) > 0): # non empty 'Target_key'
                    for tgk in o['Target_key']: # every word in 'Target_key'

                        if tgk == k: 
                            sign = True
                            opinion += 1
                            if o['Polarity'] == 'Positive':
                                pos += 1
                            else: 
                                neg += 1

            if sign == True:
                sent += 1
                if g['polar_ans']: # gold ans: True
                    t += 1
                    if g['opinions'][0]['Polarity'] == 'Positive': 
                        tp += 1
                else: # gold ans: False
                    f += 1
                    if g['opinions'][0]['Polarity'] == 'Positive': 
                        fp += 1

    total_sent2.append(sent)
    TRUE_polar_ans2.append(t)
    FALSE_polar_ans2.append(f)
    accur2.append("%.2f%%" % (100 * t/sent))
    total_opinion2.append(opinion)
    Pos2.append(pos)
    Neg2.append(neg)
    TP2.append(tp)
    FP2.append(fp)
    TN2.append(t-tp)
    FN2.append(f-fp)

df_stat2 = pd.DataFrame({'key': key2, 'total_sent': total_sent2, 'TRUE_polar_ans': TRUE_polar_ans2, 'FALSE_polar_ans': FALSE_polar_ans2,
                        'accuracy': accur2, 'TP': TP2, 'TN':  TN2, 'FP': FP2, 'FN': FN2,
                        'total_opinion': total_opinion2, 'Positive': Pos2, 'Negative': Neg2})

df_stat2 = df_stat2.sort_values(['total_sent', 'total_opinion'], ascending=False)
# df_stat2
df_stat2[df_stat2.total_sent > 2]


# total accuracy

sum_sent1 = sum(total_sent1)
sum_sent2 = sum(total_sent2)

sum_T1 = sum(TRUE_polar_ans1)
sum_T2 = sum(TRUE_polar_ans2)

print('Accuracy by target key (total_sent >= 1):')
print('rest1: ' + str(sum_T1) + '/' + str(sum_sent1) + ' = ' "%.2f%%" % (100 * sum_T1/sum_sent1))
print('rest2: ' + str(sum_T2) + '/' + str(sum_sent2) + ' = ' "%.2f%%" % (100 * sum_T2/sum_sent2))
print('total: ' + str(sum_T1+sum_T2) + '/' + str(sum_sent1+sum_sent2) + ' = ' "%.2f%%" % (100 * (sum_T1+sum_T2)/(sum_sent1+sum_sent2)))

print('-----------------------')

print('Accuracy by target key (total_sent >= 5):')
print('rest1: ' + str(96) + '/' + str(117) + ' = ' "%.2f%%" % (100 * 96/117))
print('rest2: ' + str(71) + '/' + str(79) + ' = ' "%.2f%%" % (100 * 71/79))
print('total: ' + str(96+71) + '/' + str(117+79) + ' = ' "%.2f%%" % (100 * (96+71)/(117+79)))

# df to csv
df_stat1.to_csv('../data/yelp/STAT_polar-major_rest1.csv', index=False)
df_stat2.to_csv('../data/yelp/STAT_polar-major_rest2.csv', index=False)

# combine gold1 and gold2
gold_mix = []

for i in gold1:
    gold_mix.append(i)

for j in gold2:
    gold_mix.append(j)

# output json 
with open('../data/yelp/ANS_polar-major_2rest100.json', 'w') as outfile:
    json.dump(gold_mix, outfile, indent = 4, ensure_ascii=False)

### If we only count accuracy when length of opinion is 1

key1 = list(counts1.keys())
total_sent1, TRUE_polar_ans1, FALSE_polar_ans1, accur1 = [], [], [], []
total_opinion1, Pos1, Neg1, TP1, FP1 = [], [], [], [], []
TN1, FN1 = [], []

for k in key1:
    sent, t, f = 0, 0, 0
    opinion, pos, neg, tp, fp = 0, 0, 0, 0, 0

    for g in gold1:
        sign = False
        if (len(g['opinions']) == 1): # len(['opinions']) = 1
            for o in g['opinions']: # every opinion in 'opinions'
                if (len(o['Target_key']) > 0): # non empty 'Target_key'
                    for tgk in o['Target_key']: # every word in 'Target_key'

                        if tgk == k: 
                            sign = True
                            opinion += 1
                            if o['Polarity'] == 'Positive':
                                 pos += 1
                            else: 
                                neg += 1

            if sign == True:
                sent += 1
                if g['polar_ans']: # gold ans: True
                    t += 1
                    if g['opinions'][0]['Polarity'] == 'Positive': 
                        tp += 1
                else: # gold ans: False
                    f += 1
                    if g['opinions'][0]['Polarity'] == 'Positive': 
                        fp += 1
    
    total_sent1.append(sent)
    TRUE_polar_ans1.append(t)
    FALSE_polar_ans1.append(f)

    if sent > 0:
        accur1.append("%.2f%%" % (100 * t/sent))
    else:
        accur1.append("-")

    total_opinion1.append(opinion)
    Pos1.append(pos)
    Neg1.append(neg)
    TP1.append(tp)
    FP1.append(fp)
    TN1.append(t-tp)
    FN1.append(f-fp)

# rest1's stat

df_stat1 = pd.DataFrame({'key': key1, 'total_sent': total_sent1, 'TRUE_polar_ans': TRUE_polar_ans1, 'FALSE_polar_ans': FALSE_polar_ans1,
                        'accuracy': accur1, 'TP': TP1, 'TN':  TN1, 'FP': FP1, 'FN': FN1,
                        'total_opinion': total_opinion1, 'Positive': Pos1, 'Negative': Neg1})

df_stat1 = df_stat1.sort_values(['total_sent', 'total_opinion'], ascending=False)
df_stat1[df_stat1.total_sent > 2]
# df_stat1

# rest2's statistic results by target key
# df column: key | total_sent| TRUE_polar_ans | FALSE_polar_ans | accuracy | TP | TN | FP | FN | total_opinion | Positive | Negative
# important: TP, TN, FP and FN depend on first opinion's polarity only!!!

key2 = list(counts2.keys())
total_sent2, TRUE_polar_ans2, FALSE_polar_ans2, accur2 = [], [], [], []
total_opinion2, Pos2, Neg2, TP2, FP2 = [], [], [], [], []
TN2, FN2 = [], []

for k in key2:
    sent, t, f = 0, 0, 0
    opinion, pos, neg, tp, fp = 0, 0, 0, 0, 0

    for g in gold2:
        sign = False
        if (len(g['opinions']) == 1): # non empty 'opinions'
            for o in g['opinions']: # every opinion in 'opinions'
                if (len(o['Target_key']) > 0): # non empty 'Target_key'
                    for tgk in o['Target_key']: # every word in 'Target_key'

                        if tgk == k: 
                            sign = True
                            opinion += 1
                            if o['Polarity'] == 'Positive':
                                pos += 1
                            else: 
                                neg += 1

            if sign == True:
                sent += 1
                if g['polar_ans']: # gold ans: True
                    t += 1
                    if g['opinions'][0]['Polarity'] == 'Positive': 
                        tp += 1
                else: # gold ans: False
                    f += 1
                    if g['opinions'][0]['Polarity'] == 'Positive': 
                        fp += 1

    total_sent2.append(sent)
    TRUE_polar_ans2.append(t)
    FALSE_polar_ans2.append(f)

    if sent > 0:
        accur2.append("%.2f%%" % (100 * t/sent))
    else:
        accur2.append("-")

    total_opinion2.append(opinion)
    Pos2.append(pos)
    Neg2.append(neg)
    TP2.append(tp)
    FP2.append(fp)
    TN2.append(t-tp)
    FN2.append(f-fp)

df_stat2 = pd.DataFrame({'key': key2, 'total_sent': total_sent2, 'TRUE_polar_ans': TRUE_polar_ans2, 'FALSE_polar_ans': FALSE_polar_ans2,
                        'accuracy': accur2, 'TP': TP2, 'TN':  TN2, 'FP': FP2, 'FN': FN2,
                        'total_opinion': total_opinion2, 'Positive': Pos2, 'Negative': Neg2})

df_stat2 = df_stat2.sort_values(['total_sent', 'total_opinion'], ascending=False)
# df_stat2
df_stat2[df_stat2.total_sent > 2]

# total accuracy

sum_sent1 = sum(total_sent1)
sum_sent2 = sum(total_sent2)

sum_T1 = sum(TRUE_polar_ans1)
sum_T2 = sum(TRUE_polar_ans2)

print('Accuracy by target key (total_sent >= 1):')
print('rest1: ' + str(sum_T1) + '/' + str(sum_sent1) + ' = ' "%.2f%%" % (100 * sum_T1/sum_sent1))
print('rest2: ' + str(sum_T2) + '/' + str(sum_sent2) + ' = ' "%.2f%%" % (100 * sum_T2/sum_sent2))
print('total: ' + str(sum_T1+sum_T2) + '/' + str(sum_sent1+sum_sent2) + ' = ' "%.2f%%" % (100 * (sum_T1+sum_T2)/(sum_sent1+sum_sent2)))

print('-----------------------')

print('Accuracy by target key (total_sent >= 5):')
print('rest1: ' + str(34) + '/' + str(37) + ' = ' "%.2f%%" % (100 * 96/117))
print('rest2: ' + str(41) + '/' + str(52) + ' = ' "%.2f%%" % (100 * 71/79))
print('total: ' + str(34+41) + '/' + str(37+52) + ' = ' "%.2f%%" % (100 * (34+41)/(37+52)))
