==> notebooks/38.0-new-polar-major-2rest100.ipynb <==
import pandas as pd
import json

with open('../data/yelp/ANS2-correct_polar-major_2rest100.json') as f:
    gold = json.load(f)

gold[:2]

### Rest1

gold1 = [i for i in gold if i['business_id'] == '_ab50qdWOk0DdB6XOrBitw']
key1 = []

for text in gold1:
    if len(text['opinions']) > 0:
        for op in text['opinions']:
            if len(op['Target_key']) > 0:
                for key in op['Target_key']:
                    if key not in key1:
                        key1.append(key)

len(key1)

TRUE_polar_ans1, FALSE_polar_ans1, accur1 = [], [], []
total_opinion1, Pos1, Neg1, TP1, FP1 = [], [], [], [], []
TN1, FN1 = [], []

for k in key1:
    t, f = 0, 0
    opinion, pos, neg, tp, fp = 0, 0, 0, 0, 0

    for g in gold1:
        if (len(g['opinions']) > 0): # non empty 'opinions'
            for o in g['opinions']: # every opinion in 'opinions'
                if (len(o['Target_key']) > 0): # non empty 'Target_key'
                    for tgk in o['Target_key']: # every word in 'Target_key'

                        if tgk == k: 
                            opinion += 1

                            if o['Polarity'] == 'Positive':
                                 pos += 1
                            else: 
                                neg += 1

                            if o['Polar_ans']: # gold ans: True
                                t += 1
                                if o['Polarity'] == 'Positive': 
                                    tp += 1
                            else: # gold ans: False
                                f += 1
                                if o['Polarity'] == 'Positive': 
                                    fp += 1


    total_opinion1.append(opinion)
    TRUE_polar_ans1.append(t)
    FALSE_polar_ans1.append(f)
    accur1.append("%.2f%%" % (100 * t/opinion))
    Pos1.append(pos)
    Neg1.append(neg)
    TP1.append(tp)
    FP1.append(fp)
    TN1.append(t-tp)
    FN1.append(f-fp)

# rest1's stat

df_stat1 = pd.DataFrame({'key': key1, 'total_opinion': total_opinion1, 'TRUE_polar_ans': TRUE_polar_ans1, 'FALSE_polar_ans': FALSE_polar_ans1,
                        'accuracy': accur1, 'TP': TP1, 'TN':  TN1, 'FP': FP1, 'FN': FN1,
                        'Positive': Pos1, 'Negative': Neg1})

df_stat1 = df_stat1.sort_values(['total_opinion'], ascending=False)
df_stat1[df_stat1.total_opinion >= 5]
# df_stat1

# df to csv
df_stat1.to_csv('../data/yelp/STAT2_polar-major_rest1.csv', index=False)

### Rest2

gold2 = [i for i in gold if i['business_id'] != '_ab50qdWOk0DdB6XOrBitw']
key2 = []

for text in gold2:
    if len(text['opinions']) > 0:
        for op in text['opinions']:
            if len(op['Target_key']) > 0:
                for key in op['Target_key']:
                    if key not in key2:
                        key2.append(key)

len(key2)

TRUE_polar_ans2, FALSE_polar_ans2, accur2 = [], [], []
total_opinion2, Pos2, Neg2, TP2, FP2 = [], [], [], [], []
TN2, FN2 = [], []

for k in key2:
    t, f = 0, 0
    opinion, pos, neg, tp, fp = 0, 0, 0, 0, 0

    for g in gold2:
        if (len(g['opinions']) > 0): # non empty 'opinions'
            for o in g['opinions']: # every opinion in 'opinions'
                if (len(o['Target_key']) > 0): # non empty 'Target_key'
                    for tgk in o['Target_key']: # every word in 'Target_key'

                        if tgk == k: 
                            opinion += 1

                            if o['Polarity'] == 'Positive':
                                 pos += 1
                            else: 
                                neg += 1

                            if o['Polar_ans']: # gold ans: True
                                t += 1
                                if o['Polarity'] == 'Positive': 
                                    tp += 1
                            else: # gold ans: False
                                f += 1
                                if o['Polarity'] == 'Positive': 
                                    fp += 1


    total_opinion2.append(opinion)
    TRUE_polar_ans2.append(t)
    FALSE_polar_ans2.append(f)
    accur2.append("%.2f%%" % (100 * t/opinion))
    Pos2.append(pos)
    Neg2.append(neg)
    TP2.append(tp)
    FP2.append(fp)
    TN2.append(t-tp)
    FN2.append(f-fp)

df_stat2 = pd.DataFrame({'key': key2, 'total_opinion': total_opinion2, 'TRUE_polar_ans': TRUE_polar_ans2, 'FALSE_polar_ans': FALSE_polar_ans2,
                        'accuracy': accur2, 'TP': TP2, 'TN':  TN2, 'FP': FP2, 'FN': FN2,
                        'Positive': Pos2, 'Negative': Neg2})

df_stat2 = df_stat2.sort_values(['total_opinion'], ascending=False)
df_stat2
# df_stat2[df_stat2.total_sent > 2]

# df to csv
df_stat2.to_csv('../data/yelp/STAT2_polar-major_rest2.csv', index=False)

# total accuracy

sum_sent1 = sum(total_opinion1)
sum_sent2 = sum(total_opinion2)

sum_T1 = sum(TRUE_polar_ans1)
sum_T2 = sum(TRUE_polar_ans2)

print('Accuracy by target key (total_sent >= 1):')
print('rest1: ' + str(sum_T1) + '/' + str(sum_sent1) + ' = ' "%.2f%%" % (100 * sum_T1/sum_sent1))
print('rest2: ' + str(sum_T2) + '/' + str(sum_sent2) + ' = ' "%.2f%%" % (100 * sum_T2/sum_sent2))
print('total: ' + str(sum_T1+sum_T2) + '/' + str(sum_sent1+sum_sent2) + ' = ' "%.2f%%" % (100 * (sum_T1+sum_T2)/(sum_sent1+sum_sent2)))
