==> notebooks/39.0-remove-period.ipynb <==
import pandas as pd
import json

with open('../data/yelp/RAW_yelp-food-test_2rest100.json') as f:
    raw = json.load(f)

raw[:3]

# replace '.' or '!' with ',' 
# remove '\n'

for review in range(len(raw)):
    text = raw[review]['text']
    text = text.replace('\n','')
    text_list = list(text) # because 'str' object does not support item assignment
    
    for element in range(len(text_list)):
        if element+1 < len(text_list) and element-1 >= 0:
            if text_list[element] == '.' and (text_list[element-1].isalpha() or text_list[element-1].isalnum()) and text_list[element+1] != '.':
                text_list[element] = ','

            if text_list[element] == ' ' and text_list[element+1]  == ' ' :
                text_list[element+1] = ''

            if text_list[element] == ',' and text_list[element+1] != ' ':
                text_list.insert(element+1, ' ')
    
    # print(text_list[-1])
    # text_list[-1] = '.'
    text2 = "".join(text_list)
    # print(text2)
    # print('----------')

    raw[review]['text'] = text2
    raw[review]['sent_id'] = raw[review]['review_id']
    raw[review]['opinions'] = []

raw[:5]

# output json 
with open('../data/yelp/yelp-food-noperiod_2rest100.json', 'w') as outfile:
    json.dump(raw, outfile, ensure_ascii = False, indent = 4)

with open('../data/yelp/RAW_yelp-food-test_2rest100.json') as f:
    raw2 = json.load(f)

raw2[:3]

for review in range(len(raw2)):
    text = raw2[review]['text']
    text = text.replace('\n','')

    raw2[review]['text'] = text
    raw2[review]['sent_id'] = raw2[review]['review_id']
    raw2[review]['opinions'] = []

raw2[:3]

# output json 
with open('../data/yelp/yelp-food-rawtext_2rest100.json', 'w') as outfile:
    json.dump(raw2, outfile, ensure_ascii = False, indent = 4)
