==> notebooks/40.0-infer-noperiod.ipynb <==
import pandas as pd
import json

with open('../data/yelp/inference/result/INFER_yelp-food-noperiod_2rest100.conllu.json') as f:
    infer = json.load(f)

infer[:3]

total_op = 0
nonempty_target = 0
nonempty_opinion = 0

for pred in infer:
    # print(pred['opinions'])
    if len(pred['opinions']) > 0:
        nonempty_opinion += 1
        total_op += len(pred['opinions'])

        # print(pred['text']+'\n')
        for opin in pred['opinions']:
            if len(opin['Target'][0]) > 0:
                nonempty_target += 1

                # print(opin['Target'][0], opin['Polarity'])
        # print('----------')

print(f'{nonempty_opinion} nonempty-opinion reviews in 100 reviews')
print(f'total opinions: {total_op}')
print(f'nonempty targets: {nonempty_target}')

# special cases

for pred in infer:
    if pred['sent_id'] == 'kQVI8-VqhIVq3hN2ekFCSQ' or pred['sent_id'] == 'qagu5kfTLzwXlqWYjfrS5Q':
        print(pred['text'])
        print('\n')

        for opin in pred['opinions']:
            if len(opin['Target'][0]) > 0:
                print(opin['Target'][0], opin['Polarity'])

        print('----------')



with open('../data/yelp/inference/result/INFER_yelp-food-rawtext_2rest100.conllu.json') as f:
    infer2 = json.load(f)

infer2[:3]

total_op = 0
nonempty_target = 0
nonempty_opinion = 0

for pred in infer2:
    # print(pred['opinions'])
    if len(pred['opinions']) > 0:
        nonempty_opinion += 1
        total_op += len(pred['opinions'])

        # print(pred['text']+'\n')
        for opin in pred['opinions']:
            if len(opin['Target'][0]) > 0:
                nonempty_target += 1

                # print(opin['Target'][0], opin['Polarity'])
        # print('----------')

print(f'{nonempty_opinion} nonempty-opinion reviews in 100 reviews')
print(f'total opinions: {total_op}')
print(f'nonempty targets: {nonempty_target}')

# special cases

for pred in infer2:
    if pred['sent_id'] == 'kQVI8-VqhIVq3hN2ekFCSQ' or pred['sent_id'] == 'qagu5kfTLzwXlqWYjfrS5Q':
        print(pred['text'])
        print('\n')

        for opin in pred['opinions']:
            if len(opin['Target'][0]) > 0:
                print(opin['Target'][0], opin['Polarity'])

        print('----------')
