==> notebooks/41.0-s16p-train-paragraph.ipynb <==
import pandas as pd
import json

### Combine sentences of training data

with open('../data/semEval16_plus/mixed_train.json') as f:
    sent_train = json.load(f)

sent_train[:3]

# remove noused key
for i in range(len(sent_train)):
    if sent_train[i]['dataset'] == 's16+_train':
        # print(sent_train[i])
        if len(sent_train[i]['opinions']) > 0:
            for o in range(len(sent_train[i]['opinions'])):
                del sent_train[i]['opinions'][o]['Matching']

len(sent_train)

sentid_list = []

for train in sent_train:
    sentid = ''
    if train['dataset'] == 'opener_en':
        if train['sent_id'][-2] == '-':
            sentid = train['sent_id'][:-2]
            # print(train['sent_id'])
        else:
            sentid = train['sent_id'][:-3]
    else: #s16p
        if train['sent_id'][-2] == ':':
            sentid = train['sent_id'][:-2]
            # print(train['sent_id'])
        else:
            sentid = train['sent_id'][:-3]

    if sentid not in sentid_list:
            sentid_list.append(sentid)

print(f'{len(sentid_list)} reviews in total')
# print(f'total length: {len(sent_train)}')
# print(f'nonempty opinion sent: {nonempty_sent}')        

review_train = []

for sid in sentid_list:
    review_train.append({'sent_id': sid, 'text': '', 'opinions': [], 'dataset': ''})

review_train[:5]

for i in range(len(review_train)):
    text = ''
    opinions = []
    dataset = ''

    for train in sent_train:
        sentid = ''

        if train['dataset'] == 'opener_en':
            if train['sent_id'][-2] == '-':
                sentid = train['sent_id'][:-2]
            else:
                sentid = train['sent_id'][:-3]
        else: #s16p
            if train['sent_id'][-2] == ':':
                sentid = train['sent_id'][:-2]
            else:
                sentid = train['sent_id'][:-3]
        
        
        if sentid == review_train[i]['sent_id']:
            text = text + ' ' + train['text']
            dataset = train['dataset']

            if len(train['opinions']) > 0:
                opinions = opinions + train['opinions']

    review_train[i]['text'] = text
    review_train[i]['opinions'] = opinions
    review_train[i]['dataset'] = dataset

removeindex = 0
for r in range(len(review_train)):# one element with empty opinion
    if len(review_train[r]['opinions']) == 0:
        removeindex = r
        break

review_train.remove(review_train[removeindex])
len(review_train)

for i in range(len(review_train)):
    if review_train[i]['text'][0] == ' ':
        text2 = review_train[i]['text'][1:]
        review_train[i]['text'] = text2

review_train[-1]

# output json 
with open('../data/semEval16_plus/review_train.json', 'w') as outfile:
    json.dump(review_train, outfile)

### Combine sentences of dev(test) data

with open('../data/semEval16_plus/s16+_test.json') as f:
    sent_test = json.load(f)

sent_test[:3]

# remove noused key
for i in range(len(sent_test)):
    if len(sent_test[i]['opinions']) > 0:
        for o in range(len(sent_test[i]['opinions'])):
            del sent_test[i]['opinions'][o]['Matching']

len(sent_test)

sid_list = []

for test in sent_test:
    sentid = ''
    
    if test['sent_id'][-2] == ':':
        sentid = test['sent_id'][:-2]
        # print(train['sent_id'])
    else:
        sentid = test['sent_id'][:-3]

    
    if sentid not in sid_list:
        sid_list.append(sentid)

print(f'{len(sid_list)} reviews in total (sent_test)')

review_test = []

for sid in sid_list:
    review_test.append({'sent_id': sid, 'text': '', 'opinions': [], 'dataset': ''})

review_test[:5]

for i in range(len(review_test)):
    text = ''
    opinions = []
    dataset = ''

    for test in sent_test:
        sentid = ''

        if test['sent_id'][-2] == ':':
            sentid = test['sent_id'][:-2]
        else:
            sentid = test['sent_id'][:-3]
        
        
        if sentid == review_test[i]['sent_id']:
            text = text + ' ' + test['text']
            dataset = 's16+_test'

            if len(test['opinions']) > 0:
                opinions = opinions + test['opinions']

    review_test[i]['text'] = text
    review_test[i]['opinions'] = opinions
    review_test[i]['dataset'] = dataset

for i in range(len(review_test)):
    if review_test[i]['text'][0] == ' ':
        text2 = review_test[i]['text'][1:]
        review_test[i]['text'] = text2

review_test[-1]

# output json 
with open('../data/semEval16_plus/review_test.json', 'w') as outfile:
    json.dump(review_test, outfile, ensure_ascii = False, indent = 4)
