==> notebooks/42.0-s16p-train-noperiod.ipynb <==
import pandas as pd
import json
import string

with open('../data/semEval16_plus/review_train.json') as f:
    train = json.load(f)

with open('../data/semEval16_plus/review_dev.json') as f:
    test = json.load(f)

with open('../data/yelp/inference/yelp-food-rawtext_2rest100.json') as f:
    yelp = json.load(f)

train[0]

# replace period with comma
for data in [train, test, yelp]:
    for i in range(len(data)):
        text = data[i]['text']
        text = text.replace('\n','')
        text = text.replace(' . ', ' , ')
        
        text_list = list(text)
        for element in range(len(text_list)):
            if element+1 < len(text_list) and element-1 >= 0:
                if text_list[element] == '.' and text_list[element+1] == '.':
                    continue

                if text_list[element] == '.' and text_list[element-1] == '.':
                    continue
                
                if text_list[element] == '.' and (text_list[element-1].isalpha() or text_list[element-1].isalnum()) and text_list[element+1] != '.':
                    text_list[element] = ' ,'

                if text_list[element] in string.punctuation and text_list[element+1]  == ' ' and text_list[element-1]  != ' ' :
                    text_list[element] = ' ' + text_list[element]
        
        if text_list[-1] in string.punctuation and text_list[-2] != ' ' and text_list[-2] != '.':
            text_list[-1] = ' ' + text_list[-1]

        text2 = "".join(text_list)
        text2 = text2.replace('  ', ' ')
        data[i]['text'] = text2


for i in train[20:30]:
    print(i['text'])
    print('------------')

def lower_exceptI (txt):
    if txt == 'i' or txt == 'I':
        return 'I'
    else:
        lst = list(txt.lower())
        for i in range(len(lst)):
            if i+1 < len(lst) and i-1 >= 0:
                if lst[i] == 'i' and (lst[i+1] == ' ' or lst[i+1] in string.punctuation) and lst[i-1] == ' ':
                    lst[i] = lst[i].upper()
                elif lst[i] == 'i' and (lst[i+1].isalpha() == False) and (lst[i-1].isalpha() == False):
                    lst[i] = lst[i].upper()
            elif i == 0 and lst[i] == 'i' and (lst[i+1] == ' ' or lst[i+1] in string.punctuation):
                lst[i] = lst[i].upper()
            else:
                lst[i] = lst[i].lower()
        return ("".join(lst))

# convert uppercase to lowercase for all letters (also for targets and expression)
# only I be uppercase
for data in [train, test, yelp]:
    for i in range(len(data)):
        data[i]['text'] = lower_exceptI(data[i]['text'])

        if len(data[i]['opinions']) > 0:
            for op in range(len(data[i]['opinions'])):
                source = data[i]['opinions'][op]['Source'][0]
                target = data[i]['opinions'][op]['Target'][0]
                exp = data[i]['opinions'][op]['Polar_expression'][0]

                source_lower = [lower_exceptI(s) for s in source]
                target_lower = [lower_exceptI(t) for t in target]
                exp_lower = [lower_exceptI(e) for e in exp]

                data[i]['opinions'][op]['Source'][0] = source_lower
                data[i]['opinions'][op]['Target'][0] = target_lower
                data[i]['opinions'][op]['Polar_expression'][0] = exp_lower 

# replace , ! with and
for data in [train, test, yelp]:
    for i in range(len(data)):
        text = data[i]['text']
        text = text.replace(' , ', ' and ')
        text = text.replace(' ! ', ' and ')
        text = text.replace(' and !', ' ! !')
        text = text.replace('and but', 'but')
        data[i]['text'] = text

for i in train[20:30]:
    print(i['text'])
    print('------------')

train[-7]

yelp[1]

# output json 
with open('../data/semEval16_plus/noperiod_dev.json', 'w') as outfile1:
    json.dump(test, outfile1, ensure_ascii = False, indent = 4)

# output json 
with open('../data/semEval16_plus/noperiod_train.json', 'w') as outfile2:
    json.dump(train, outfile2, ensure_ascii = False, indent = 4)

# output json 
with open('../data/yelp/inference/yelp-food-nopr_2rest100.json', 'w') as outfile2:
    json.dump(yelp, outfile2, ensure_ascii = False, indent = 4)
